Added manual garbage collection
Trying to fix an issue with high memory usage, especially when there are issues with connections.
Currently, collection happens every 90s.
This could be significantly reduced and triggered specifically when issues cause more allocations.
There is also a way to tell the runtime to limit the max heap used so it handles it on its own.
We will see what will be used.